backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "easyocr>=1.7.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "groq>=0.32.0",
    "pillow>=11.3.0",
    "streamlit>=1.50.0",
    "torch>=2.8.0",
    "transformers>=4.57.0",
]
